// src/types/analytics/call-type.ts
/*
 * Copyright 2021 ALE International
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy of this
 * software and associated documentation files (the "Software"), to deal in the Software
 * without restriction, including without limitation the rights to use, copy, modify, merge,
 * publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons
 * to whom the Software is furnished to do so, subject to the following conditions:
 *
 * The above copyright notice and this permission notice shall be included in all copies or
 * substantial portions of the Software.
 *
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING
 * BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
 * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
 * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */

/**
 * CallType represents the possible call types in a charging ticket.
 */
export enum CallType {
    PublicNetworkCall = 'PublicNetworkCall',
    PublicNetworkCallThroughPrivateNetwork = 'PublicNetworkCallThroughPrivateNetwork',
    PrivateNetworkCall = 'PrivateNetworkCall',
    LocalNetworkCall = 'Local etworkCall',
    PublicNetworkIncomingCall = 'PublicNetwork IncomingCall',
    PublicNetworkIncomingCallThroughPrivateNetwork = 'PublicNetworkIncomingCallThroughPrivateNetwork',
    Unspecified = 'Unspecified',
    PrivateNetworkOutgoingCallToPublicNetwork = 'PrivateNetworkOutgoingCallToPublicNetwork',
    PrivateNetworkOutgoingCallToPrivateNetwork = 'PrivateNetworkOutgoingCallToPrivateNetwork',
    PublicNetworkIncomingCallToPrivateNetwork = 'PublicNetworkIncomingCallToPrivateNetwork',
    PrivateNetworkIncomingCallToPrivateNetwork = 'PrivateNetworkIncomingCallToPrivateNetwork',
    PublicOrPrivateNetworkOutgoingCallThroughPrivateNetwork = 'PublicOrPrivateNetworkOutgoingCallThroughPrivateNetwork',
    PublicOrPrivateNetworkIncomingCallThroughPrivateNetwork = 'PublicOrPrivateNetworkIncomingCallThroughPrivateNetwork',
    PrivateNetworkIncomingCall = 'PrivateNetworkIncomingCall',
    LocalNode = 'LocalNode',
    LocalTransit = 'LocalTransit',
}

export namespace CallType {
    export function isCallType(value: string): value is CallType {
        return Object.values(CallType).includes(value as CallType);
    }
}